Refactor: Convert ListView to RecyclerView in LibraryFragment
