+++ app/src/main/java/org/kiwix/kiwixmobile/views/LanguageSelectDialog.java
@@ -30,7 +30,7 @@ import android.widget.ListView;
 import android.widget.TextView;
 
 import org.kiwix.kiwixmobile.R;
-import org.kiwix.kiwixmobile.library.LibraryAdapter;
+import org.kiwix.kiwixmobile.library.LibraryRecyclerViewAdapter;
 import org.kiwix.kiwixmobile.utils.LanguageUtils;
 
 import java.util.List;
@@ -46,7 +46,7 @@ public class LanguageSelectDialog extends AlertDialog {
   }
 
   public static class Builder extends AlertDialog.Builder {
-    private List<LibraryAdapter.Language> languages;
+    private List<LibraryRecyclerViewAdapter.Language> languages;
     private Map<String, Integer> languageCounts;
     private boolean singleSelect = false;
     private String selectedLanguage;
@@ -60,7 +60,7 @@ public class LanguageSelectDialog extends AlertDialog {
       super(context, themeResId);
     }
 
-    public Builder setLanguages(List<LibraryAdapter.Language> languages) {
+    public Builder setLanguages(List<LibraryRecyclerViewAdapter.Language> languages) {
       this.languages = languages;
       return this;
     }
@@ -114,13 +114,13 @@ public class LanguageSelectDialog extends AlertDialog {
     }
   }
 
-  private static class LanguageArrayAdapter extends ArrayAdapter<LibraryAdapter.Language> {
+  private static class LanguageArrayAdapter extends ArrayAdapter<LibraryRecyclerViewAdapter.Language> {
     private Map<String, Integer> languageCounts;
     private Context context;
     private boolean singleSelect;
     private String selectedLanguage;
 
-    public LanguageArrayAdapter(Context context, int textViewResourceId, List<LibraryAdapter.Language> languages,
+    public LanguageArrayAdapter(Context context, int textViewResourceId, List<LibraryRecyclerViewAdapter.Language> languages,
                                 Map<String, Integer> languageCounts, boolean singleSelect, String selectedLanguage) {
       super(context, textViewResourceId, languages);
       this.languageCounts = languageCounts;
@@ -152,7 +152,7 @@ public class LanguageSelectDialog extends AlertDialog {
       holder.checkBox
           .setOnCheckedChangeListener((compoundButton, b) -> getItem(position).active = b);
 
-      LibraryAdapter.Language language = getItem(position);
+      LibraryRecyclerViewAdapter.Language language = getItem(position);
       holder.language.setText(language.language);
       holder.languageLocalized.setText(context.getString(R.string.language_localized,
               language.languageLocalized));
