+++ app/src/main/res/layout/activity_library.xml
@@ -47,7 +47,7 @@
       android:layout_width="wrap_content"
       android:layout_height="wrap_content">
 
-      <ListView
+      <android.support.v7.widget.RecyclerView
         android:id="@+id/library_list"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
